native modules chunks changes

+++ src/features/MeetingScreens/MeetingScreen/index.tsx
@@ -24,10 +24,10 @@ import { useMeetingAccess } from "src/hooks/useMeetingAccess"
 import { useAudioRecorder } from "src/hooks/useAudioRecorder"
 import { NativeEventEmitter, NativeModules } from "react-native"
 const { ScreenRecorder } = NativeModules
-import RNFS from 'react-native-fs';
-import Base64 from 'react-native-base64';
+import RNFS from "react-native-fs"
+import Base64 from "react-native-base64"
 
-const recordingUrl = Config.SOCKET_RECORDING_URL
+// const recordingUrl = Config.SOCKET_RECORDING_URL
 
 type ParamList = {
   Detail: {
@@ -84,13 +84,13 @@ const MeetingScreen = () => {
     rtcError,
   } = useWebRtc(instanceMeetingOwner!)
   const { goBack } = useNavigation()
-  const [isRecording, setIsRecording] = useState(false);
+  const [isRecording, setIsRecording] = useState(false)
 
   const [invitedParticipants, setInvitedParticipants] = useState<any[]>([])
   const [meInvited, setMeInvited] = useState<boolean | null>(null)
 
-  const recordingNameRef = useRef<any>();
-  const startTimeRef = useRef<any>();
+  const recordingNameRef = useRef<any>()
+  const startTimeRef = useRef<any>()
 
   const {
     joinEvent,
@@ -143,44 +143,46 @@ const MeetingScreen = () => {
     sheetParticipantsRef.current?.open()
   }
 
-
   const saveChunkToFile = async (chunk: any) => {
+    console.log("saveChunkToFile")
     try {
-      const filePath = `${RNFS.DocumentDirectoryPath}/recording-${Date.now()}.webm`;
-      const decodedData = Base64.decode(chunk);
-      await RNFS.writeFile(filePath, decodedData, 'base64');
+      const filePath = `${
+        RNFS.DocumentDirectoryPath
+      }/recording-${roomId}.webm`
+      const decodedData = Base64.decode(chunk)
+      const fileExists = await RNFS.exists(filePath)
+      if (fileExists) {
+        // Append to file instead of overwriting
+        await RNFS.appendFile(filePath, decodedData, "base64")
+      } else {
+        // If file doesn't exist, create a new file
+        await RNFS.writeFile(filePath, decodedData, "base64")
+      }
 
-      RNFS.stat(filePath)
-        .then((statResult) => {
-          console.log('File size:', statResult.size);
-          if (statResult.size > 0) {
-            RNFS.readFile(filePath, 'base64')
-              .then((fileData) => {
-                console.log('File data (base64):', fileData);
-              })
-              .catch((error) => {
-                console.error('Failed to read file:', error);
-              });
-          } else {
-            console.log('File is empty.');
-          }
-        })
-        .catch((error) => {
-          console.error('Failed to get file stats:', error);
-        });
+    // Get file stats to check the size
+    const statResult = await RNFS.stat(filePath);
+    console.log("File size:", statResult.size);
 
-      console.log(`Chunk saved to file at: ${filePath}`);
-    } catch (error) {
-      console.error("Failed to save chunk to file:", error);
+    // Optionally, read the file after saving (for debugging)
+    if (statResult.size > 0) {
+      const fileData = await RNFS.readFile(filePath, 'base64');
+      console.log("File data (base64):", fileData);
+    } else {
+      console.log("File is empty.");
     }
-  };
+
+      console.log(`Chunk saved to file at: ${filePath}`)
+    } catch (error) {
+      console.error("Failed to save chunk to file:", error)
+    }
+  }
 
   useEffect(() => {
     const eventEmitter = new NativeEventEmitter(ScreenRecorder)
     const chunkListener = eventEmitter.addListener(
       "onVideoChunk",
       ({ chunk }) => {
-        console.log("Received chunk:", chunk)
+        // console.log("Received chunk:", chunk)
         sendChunkToServer(chunk)
       }
     )
@@ -191,11 +193,12 @@ const MeetingScreen = () => {
   }, [])
 
   const startRecording = async () => {
+    console.log("startRecordingstartRecordingstartRecording")
     try {
       // if (wsRef.current?.readyState === WebSocket.OPEN) {
       //  recordingNameRef.current = `recording-mobile-${Date.now()}`
       // startTimeRef.current = Date.now()
-      await ScreenRecorder.startRecording();
+      await ScreenRecorder.startRecording()
       console.log("Recording started")
       setIsRecording(true)
       // }
@@ -235,48 +238,48 @@ const MeetingScreen = () => {
     }
   }
 
-  useEffect(() => {
-    let reconnectTimeout: NodeJS.Timeout | null = null
+  // useEffect(() => {
+  //   let reconnectTimeout: NodeJS.Timeout | null = null
 
-    const connectWebSocket = () => {
-      if (wsRef.current) {
-        wsRef.current.close()
-      }
+  //   const connectWebSocket = () => {
+  //     if (wsRef.current) {
+  //       wsRef.current.close()
+  //     }
 
-      wsRef.current = new WebSocket(recordingUrl!)
+  //     wsRef.current = new WebSocket(recordingUrl!)
 
-      wsRef.current.onopen = () => {
-        console.log("WebSocket connected")
-        if (reconnectTimeout) {
-          clearTimeout(reconnectTimeout)
-          reconnectTimeout = null
-        }
-      }
+  //     wsRef.current.onopen = () => {
+  //       console.log("WebSocket connected")
+  //       if (reconnectTimeout) {
+  //         clearTimeout(reconnectTimeout)
+  //         reconnectTimeout = null
+  //       }
+  //     }
 
-      wsRef.current.onerror = (error) => {
-        console.error("WebSocket error:", error)
-        console.log("WebSocket connection error. Reconnecting...")
-        if (!reconnectTimeout) {
-          reconnectTimeout = setTimeout(connectWebSocket, 3000)
-        }
-      }
+  //     wsRef.current.onerror = (error) => {
+  //       console.error("WebSocket error:", error)
+  //       console.log("WebSocket connection error. Reconnecting...")
+  //       if (!reconnectTimeout) {
+  //         reconnectTimeout = setTimeout(connectWebSocket, 3000)
+  //       }
+  //     }
 
-      wsRef.current.onclose = () => {
-        console.warn("WebSocket connection closed")
-      }
-    }
+  //     wsRef.current.onclose = () => {
+  //       console.warn("WebSocket connection closed")
+  //     }
+  //   }
 
-    connectWebSocket()
+  //   connectWebSocket()
 
-    return () => {
-      if (reconnectTimeout) clearTimeout(reconnectTimeout)
-      wsRef.current?.close()
-    }
-  }, [])
+  //   return () => {
+  //     if (reconnectTimeout) clearTimeout(reconnectTimeout)
+  //     wsRef.current?.close()
+  //   }
+  // }, [])
 
   const sendChunkToServer = async (base64Chunk: any) => {
     try {
-      await saveChunkToFile(base64Chunk);
+      await saveChunkToFile(base64Chunk)
       // if (wsRef.current?.readyState === WebSocket.OPEN) {
       //   wsRef.current.send(
       //     JSON.stringify({
@@ -284,7 +287,7 @@ const MeetingScreen = () => {
       //       fileExtension: "mp4",
       //       chunks: base64Chunk,
       //       action: "stream",
-      //     })          
+      //     })
       //   )
       //   console.log('chunk sent');
       // }
@@ -369,20 +372,23 @@ const MeetingScreen = () => {
     },
   ]
 
+  //   console.log(eventId, "eventIdeventIdeventId")
+  //
+
   useEffect(() => {
     if (!!participants.length && socketInstance) {
       joinEvent({ eventId: String(eventId) })
     }
   }, [participants?.length, socketInstance])
 
-  if (rtcError) {
-    Toast.show({
-      type: "error",
-      text1:
-        "Connection to media servers cannot be established, please consider rejoining",
-    })
-    goBack()
-  }
+  // if (rtcError) {
+  //   Toast.show({
+  //     type: "error",
+  //     text1:
+  //       "Connection to media servers cannot be established, please consider rejoining",
+  //   })
+  //   goBack()
+  // }
   // if (!participants?.length) {
   //   return <Loading />
   // }
